add super.viewWillAppear(animated)
When you override a method from a superclass, it is always important to
call the corresponding super method.
Even if the super class is not doing anything with it today, Apple
might one day change the implementation and your code will mysteriously
stop working. If you really don't need to do anything in that method,
leave it out of your code entirely, and the super's method will run as
usual, without any intervention on your part.